edited FavoriteMeal model and changed corresponding repository,service, serviceImpl and controller.

+++ src/main/java/com/chompfooddeliveryapp/service/serviceImpl/FavoriteMenuServiceImpl.java
@@ -1,7 +1,7 @@
 package com.chompfooddeliveryapp.service.serviceImpl;
 
 import com.chompfooddeliveryapp.exception.FavoriteNotFoundException;
-import com.chompfooddeliveryapp.model.enums.MenuCategory;
+import com.chompfooddeliveryapp.exception.MenuNotFoundException;
 import com.chompfooddeliveryapp.model.meals.FavoriteMeal;
 import com.chompfooddeliveryapp.model.meals.MenuItem;
 import com.chompfooddeliveryapp.model.users.User;
@@ -13,42 +13,61 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FavoriteMenuServiceImpl implements FavoriteMealService {
 
     private final FavoriteMealRepository favoriteMealRepository;
-
-
+    private final UserRepository userRepository;
+    private final MenuItemRepository menuItemRepository;
 
     @Autowired
-    public FavoriteMenuServiceImpl(FavoriteMealRepository favoriteMealRepository, MenuItemRepository menuItemRepository, UserRepository userRepository) {
+    public FavoriteMenuServiceImpl(FavoriteMealRepository favoriteMealRepository, UserRepository userRepository, MenuItemRepository menuItemRepository) {
         this.favoriteMealRepository = favoriteMealRepository;
+        this.userRepository = userRepository;
+        this.menuItemRepository = menuItemRepository;
     }
+
 
     @Override
     public FavoriteMeal createFavoriteMeal(Long userId, Long menuId) {
+        Optional<User> user = userRepository.findUserById(userId);
+        Optional<MenuItem> menuItem = menuItemRepository.findMenuItemById(menuId);
+        if (user.isEmpty() || menuItem.isEmpty()) {
+            throw new MenuNotFoundException("User or MenuItem not found");
+        }
 
-        FavoriteMeal newfavoriteMeal;
-        FavoriteMeal favoriteMeal = favoriteMealRepository.findFavoriteMealByUser_idAndMenu_id(userId, menuId)
-                .orElse(newfavoriteMeal = new FavoriteMeal(userId, menuId));
+        Optional<FavoriteMeal> favoriteMeal = favoriteMealRepository.findFavoriteMealByUseridAndMenuid(user.get(), menuItem.get());
+        if (favoriteMeal.isPresent()) {
+            throw new MenuNotFoundException("Item is already a favorite");
+        };
+        FavoriteMeal newfavoriteMeal = new FavoriteMeal(user.get(), menuItem.get());
         return favoriteMealRepository.save(newfavoriteMeal);
     }
 
     @Override
     public String removeFromFavoriteMeal(Long userId, Long menuId) {
-        FavoriteMeal favoriteMeal = favoriteMealRepository.findFavoriteMealByUser_idAndMenu_id(userId, menuId)
+        Optional<User> user = userRepository.findUserById(userId);
+        Optional<MenuItem> menuItem = menuItemRepository.findMenuItemById(menuId);
+        if (user.isEmpty() || menuItem.isEmpty()) throw new MenuNotFoundException("User or MenuItem not found");
+        FavoriteMeal favoriteMeal = favoriteMealRepository.findFavoriteMealByUseridAndMenuid(user.get(), menuItem.get())
                 .orElseThrow(()->new FavoriteNotFoundException("favorite meal with id: " + menuId + " does not exist"));
         favoriteMealRepository.delete(favoriteMeal);
         return "favorite meal with " + menuId + " has been removed from favorite.";
     }
 
     public List<FavoriteMeal> getAllFavoriteMealsByAUser(Long userId) {
-        return favoriteMealRepository.findByUser_id(userId);
+        Optional<User> user = userRepository.findUserById(userId);
+        if (user.isEmpty()) throw new MenuNotFoundException("User or MenuItem not found");
+        return favoriteMealRepository.findByUserid(user.get());
     }
 
     public FavoriteMeal getAParticularFavoriteMeal(Long userId, Long menuId){
-       return favoriteMealRepository.findFavoriteMealByUser_idAndMenu_id(userId, menuId)
+        Optional<User> user = userRepository.findUserById(userId);
+        Optional<MenuItem> menuItem = menuItemRepository.findMenuItemById(menuId);
+        if (user.isEmpty() || menuItem.isEmpty()) throw new MenuNotFoundException("User or MenuItem not found");
+       return favoriteMealRepository.findFavoriteMealByUseridAndMenuid(user.get(), menuItem.get())
                 .orElseThrow( ()-> new FavoriteNotFoundException("Favorite meal does not exist"));
     }
